Windows Phone Mentioned in Docs
Proposal for removing lines:
-field PickerReturned:1000
Windows Phone only. The app was activated after the completion of a picker.
-field WalletAction:1001
Windows Phone only. The app was activated to perform a Wallet operation.
-field PickFileContinuation:1002
Windows Phone only. The app was activated after the app was suspended for a file picker operation.
-field PickSaveFileContinuation:1003
Windows Phone only. The app was activated after the app was suspended for a file save picker operation.
-field PickFolderContinuation:1004
Windows Phone only. The app was activated after the app was suspended for a folder picker operation.
-field WebAuthenticationBrokerContinuation:1005
Windows Phone only. The app was activated after the app was suspended for a web authentication broker operation.